--- src/js/library.js
@@ -1,7 +1,0 @@
-import { refs } from "./refs"
-export function createLibrary() {
-    refs.filmsContainerRef.innerHTML = "Это библиотека!"
-    refs.headerButtons.innerHTML = `
-        <button type="button">Watched</button>
-        <button type="button">queue</button>`
-}
