+++ src/partials/Lyuda.html
@@ -1,4 +1,4 @@
-<div>
+<!-- <div>
 <h1>This is Lyuda`s partial</h1>
 <form class="search-form" id="search-form">
         <input
@@ -14,4 +14,4 @@
     </header>
 
 </div>
-
+ -->
