+++ src/partials/header/header.html
@@ -1,6 +1,6 @@
 <div class="header">
   <ul>
-    <li><a href="index.html">home</a></li>
+    <li><a class="page-home" href="#">home</a></li>
     <li><a class="page-library" href="#">my library</a></li>
   </ul>
   <div class="header-functional">
